requirements.txt
python-telegram-bot>=21.0
supabase==2.3.4
openai==1.12.0
pandas==2.3.2
openpyxl==3.1.5
python-dotenv==1.0.0
Pillow>=10.0.0
requests==2.31.0
aiohttp>=3.9.0
